Refactor and fix pi hw implementation
